#162 CR fixes
